app: move update route  to codyapp package
